fix: moving query for leagues to query file
